Index.js brokedown to modules
